dt.count

A function that counts the number of rows that meets the qualification such as filter and the by. If both filter and the by are not set up, then it will return the number of rows.

dt.count(
  dat,
  the.filter = NULL,
  grouping.variables = NULL,
  grouping.type = "keyby",
  count.name = "N"
)

Arguments

dat
a data.frame object.

the.filter
a character value, logical value, or expression stating the logical operations to be performed in filtering the data prior to calculating the.function. Defaults to "NULL" unless specified.

grouping.variables
a character vector specifying variables to group by in performing the computation. Only values that exist in names(dat) will be used.

grouping.type
a character value specifying whether the grouping should be sorted (keyby) or as is (by). Defaults to keyby unless "by" is specified.

count.name
a character value specifying the name of the column of counts in the resulting table. Defaults to "N" unless specified.

Source

create.filter.expression.R

Note

the data.frame dat will be converted to a data.table object to facilitate adding the new column by reference (e.g. efficiently with regard to the usage of memory)

Examples

 age.name = "Age"
 income.name = "Income"
 region.name = "Region"
 gender.name = "Gender"

 dt.count(dat = formulaic::snack.dat)

#>        N
#> 1: 23000

 dt.count(dat = formulaic::snack.dat,
 grouping.variables = c(region.name, gender.name))

#>       Region Gender    N
#> 1:   Midwest Female 2139
#> 2:   Midwest   Male 2093
#> 3: Northeast Female 3588
#> 4: Northeast   Male 3335
#> 5:     South Female 1679
#> 6:     South   Male 1541
#> 7:      West Female 4278
#> 8:      West   Male 4347

 dt.count(dat = formulaic::snack.dat,
 the.filter = "Age > 65",
 grouping.variables = c(region.name, gender.name),
 count.name = "Records with Age > 65")

#>       Region Gender Records with Age > 65
#> 1:   Midwest Female                   736
#> 2:   Midwest   Male                   897
#> 3: Northeast Female                  1403
#> 4: Northeast   Male                  1265
#> 5:     South Female                   575
#> 6:     South   Male                   483
#> 7:      West Female                  1564
#> 8:      West   Male                  1518

 dt.count(dat = formulaic::snack.dat,
 the.filter = 'get(age.name) > 20',
 grouping.variables = c(age.name,income.name),
 grouping.type = 'keyby', count.name = "counts")

#> Error in .checkTypos(e, names_x): Object 'age.name' not found amongst Age, Gender, Income, Region, Persona and 20 more